+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * `KYCRecovery` reviewable request ingestion
 * `GetKYCRecoveryRequests` to handle `/v3/kyc_recovery_requests`, `/v3/kyc_recovery_requests/{id}`
 * Docs on kyc recovery
+* `kyc_data` relationship to `account` resource type
+* `kyc_data` include parameter to `/v3/accounts/{id}`
 * `RemoveAssetPairOp` operation ingestion
 * Tests for operation details handling
 * `/v3/accounts/{id}/sales/{sale_id}` endpoint to get sale by id, if account is allowed to participate